backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-socketio[asyncio_client]==5.11.4
python-multipart==0.0.9
httpx==0.27.2
aiohttp==3.10.10
pydantic==2.9.2
pydantic-settings==2.5.2
PyYAML==6.0.2
aiosqlite==0.20.0
elasticsearch[async]==8.15.1
apscheduler==3.10.4
python-nmap==0.7.1
python-dotenv==1.0.1
python-gvm==24.8.0
asyncssh==2.17.0
docker==7.1.0
